+++ main/BubbleSort.h
@@ -1,3 +1,3 @@
 #pragma once
 #include "DataGenerator.h"
-void BubbleSort(int* arr, int n, int& count_compare);
+void BubbleSort(int* arr, int n, double& count_compare);
