+++ main/FlashSort.h
@@ -1,5 +1,5 @@
 #pragma once
 #include "InsertionSort.h"
 #include "DataGenerator.h"
-void FlashSort(int* arr, int n, int& count_compare);
+void FlashSort(int* arr, int n, double& count_compare);
