+++ main/HeapSort.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "DataGenerator.h"
-void MaxHeapify(int* arr, int n, int i, int& count_compare);
-void BuildMaxHeap(int* arr, int n, int& count_compare);
-void HeapSort(int* arr, int n, int& count_compare);
+void MaxHeapify(int* arr, int n, int i, double& count_compare);
+void BuildMaxHeap(int* arr, int n, double& count_compare);
+void HeapSort(int* arr, int n, double& count_compare);
