+++ main/InsertionSort.h
@@ -1,2 +1,2 @@
 #pragma once
-void InsertionSort(int* arr, int n, int& count_compare);
+void InsertionSort(int* arr, int n, double& count_compare);
