+++ main/MergeSort.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <iostream>
-void Merge(int* arr, int left, int mid, int right, int& count_compare);
-void MergeSort(int* arr, int left, int right, int& count_compare);
+void Merge(int* arr, int left, int mid, int right, double& count_compare);
+void MergeSort(int* arr, int left, int right, double& count_compare);
