+++ main/QuickSort.h
@@ -1,4 +1,4 @@
 #pragma once
 #include "DataGenerator.h"
-int Partition(int* arr, int left, int right, int& count_compare);
-void QuickSort(int* arr, int left, int right, int& count_compare);
+int Partition(int* arr, int left, int right, double& count_compare);
+void QuickSort(int* arr, int left, int right, double& count_compare);
