+++ main/RadixSort.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <iostream>
-void CountingSortForRadix(int* arr, int n, int div, int& count_compare);
-void RadixSort(int* arr, int n, int& count_compare);
+void CountingSortForRadix(int* arr, int n, int div, double& count_compare);
+void RadixSort(int* arr, int n, double& count_compare);
