+++ main/ShakerSort.h
@@ -1,4 +1,4 @@
 #pragma once
 #include "DataGenerator.h"
-void ShakerSort(int* arr, int n, int& count_compare);
+void ShakerSort(int* arr, int n, double& count_compare);
