+++ main/ShellSort.h
@@ -1,4 +1,4 @@
 #pragma once
-void ShellSort(int* arr, int n, int& count_compare);
+void ShellSort(int* arr, int n, double& count_compare);
